g4-common: remove qcom iZat
closes https://github.com/Suicide-Squirrel/issues_pie/issues/33

iZat is qualcomm's way of providing in-house location services
(i.e. without gps satellites) by using sensors and WiFi etc.
It is aimed to be using less power but there are also reports on
that it consumes more. Other then that the whole thing makes sense
only when you have apps and map material which providing in-house
"routing".
The idea was to use this in big malls and such but there are just
a few (japan/korea) ones and so the whole thing is IMHO completely
useless (as of now).